# contracts/bff.openapi.yaml
openapi: 3.0.3
info:
  title: BFF API
  version: 0.2.0
  description: Public-facing API for the frontend. Producer - bff (TypeScript). Consumer - frontend. Uses camelCase.
servers:
  - url: http://localhost:4000
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /api/realtime/session:
    post:
      summary: Create realtime session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RealtimeSessionCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RealtimeSessionResponse"
        "400":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
  /api/alerts:
    post:
      summary: Create alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "400":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
    get:
      summary: List alerts
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertListItem"
        "502":
          $ref: "#/components/responses/Error"
components:
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status: { type: string, example: ok }
    RealtimeSessionCreate:
      type: object
      required: [sessionId, language, alertPhrases, emergencyPhrases]
      properties:
        sessionId:
          type: string
          description: Frontend-generated call/session UUID.
        language:
          type: string
          enum: [pl, en]
          description: UI and transcription language.
        alertPhrases:
          type: array
          items:
            type: string
          description: Level 1 code phrases.
        emergencyPhrases:
          type: array
          items:
            type: string
          description: Level 2 code phrases.
    RealtimeSessionResponse:
      type: object
      required: [clientSecret, expiresAt, model]
      properties:
        clientSecret:
          type: string
          description: Ephemeral OpenAI client secret.
        expiresAt:
          type: string
          format: date-time
          description: Secret expiry timestamp.
        model:
          type: string
          description: Realtime model name.
    AlertCreate:
      type: object
      required: [sessionId, level, source, language, recipientName, recipientPhone, location]
      properties:
        sessionId:
          type: string
          description: Call/session UUID.
        level:
          type: string
          enum: [alert, emergency]
          description: Alert severity.
        source:
          type: string
          enum: [keyword, agent, manual]
          description: Alert origin.
        triggerPhrase:
          type: string
          description: Detected phrase or manual trigger note.
        language:
          type: string
          enum: [pl, en]
          description: Alert language.
        recipientName:
          type: string
          description: Trusted contact name.
        recipientPhone:
          type: string
          description: Trusted contact phone.
        location:
          oneOf:
            - $ref: "#/components/schemas/Location"
            - type: "null"
          description: GPS coordinates or null if unavailable.
        transcriptSnippet:
          type: string
          description: Short transcript excerpt.
    Location:
      type: object
      required: [latitude, longitude, accuracy]
      properties:
        latitude:
          type: number
          description: GPS latitude.
        longitude:
          type: number
          description: GPS longitude.
        accuracy:
          type: number
          description: GPS accuracy in meters.
    Delivery:
      type: object
      required: [channel, status]
      properties:
        channel:
          type: string
          enum: [mock, sms, voice, telegram]
          description: Delivery channel.
        status:
          type: string
          enum: [sent, failed]
          description: Delivery status.
    Alert:
      type: object
      required: [id, mapsUrl, deliveries, createdAt]
      properties:
        id:
          type: string
          description: Alert id.
        mapsUrl:
          type: string
          description: Google Maps link.
        deliveries:
          type: array
          items:
            $ref: "#/components/schemas/Delivery"
          description: Delivery results.
        createdAt:
          type: string
          format: date-time
          description: Alert creation time.
    AlertListItem:
      type: object
      required: [id, sessionId, createdAt, level, source, location, deliveries]
      properties:
        id:
          type: string
          description: Alert id.
        sessionId:
          type: string
          description: Call/session UUID.
        createdAt:
          type: string
          format: date-time
          description: Alert creation time.
        level:
          type: string
          enum: [alert, emergency]
          description: Alert severity.
        source:
          type: string
          enum: [keyword, agent, manual]
          description: Alert origin.
        triggerPhrase:
          type: string
          description: Detected phrase or manual trigger note.
        transcriptSnippet:
          type: string
          description: Short transcript excerpt.
        location:
          oneOf:
            - $ref: "#/components/schemas/Location"
            - type: "null"
          description: GPS coordinates or null if unavailable.
        deliveries:
          type: array
          items:
            $ref: "#/components/schemas/Delivery"
          description: Delivery results.
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, UPSTREAM_ERROR, INTERNAL_ERROR]
            message:
              type: string
            details:
              type: object